Split off error retext config from more noisy warning config.
