+++ .remarkrc.warning
@@ -1,5 +1,5 @@
 {
   "plugins": [
-      "./.retext"
+      "./.retext.warning"
     ]
 }
